Coursework 1 of the project
Coursework 1 of the project
The report will provide detailed information and analysis of the FASAM system by providing clear, detailed functional and non-functional requirements, partial specification that outlines the key features and functionalities of the system, architectural design using the UML Modelling Technique, Prototype for the Proposed System User Interface and Requirements Validation and Management Strategy. Therefore, the report will provide a detailed solution of a system that solves this real time, large scale and safety concerning issue.